refactor: 체스판 위치 관리하는 기능 Position 구현 및 Square에 기물 정보 저장
