+++ src/main/java/chess/model/board/Position.java
@@ -1,0 +1,12 @@
+package chess.model.board;
+
+public class Position {
+
+    private final Rank rank;
+    private final File file;
+
+    public Position(final Rank rank, final File file) {
+        this.rank = rank;
+        this.file = file;
+    }
+}
